:art: Move App class methods into modules
