+++ app_functionality/ui_formating.rb
@@ -1,0 +1,14 @@
+module UiFormat
+  def select_option
+    puts ''
+    print 'Select option'
+    puts ''
+  end
+
+  def back_to_menu
+    puts ''
+    print 'Press Enter to go in menu'
+    gets.chomp
+    puts ''
+  end
+end
